requirements.txt
numpy==1.19.4
opencv-python==4.4.0.46
Pillow<8
pygame==2.0.0
PyOpenGL==3.1.5
